+++ Deeplex.Saverwalter.Model/model/VertragVersion.cs
@@ -27,11 +27,11 @@ namespace Deeplex.Saverwalter.Model
         [Required]
         public DateOnly Beginn { get; set; }
         [Required]
-        public double Grundmiete { get; set; }
+        public decimal Grundmiete { get; set; }
         public string? Notiz { get; set; }
         public DateTime CreatedAt { get; private set; }
         public DateTime LastModified { get; set; }
-        public VertragVersion(DateOnly beginn, double grundmiete, int personenzahl)
+        public VertragVersion(DateOnly beginn, decimal grundmiete, int personenzahl)
         {
             Beginn = beginn;
             Grundmiete = grundmiete;
